Main Configs zu Config; GCAM fixes;
