+++ addons/config/configs/CBA_Stuff.hpp
@@ -4,4 +4,4 @@
 */
 cba_settings_whitelist[] = {"76561198029318101", "admin"};
 enableTargetDebug = 1;
-enableDebugConsole[] = {"76561198029318101"};
+enableDebugConsole[] = {"76561198029318101", "76561198127241859"};
