+++ addons/config/configs/CfgCommands.hpp
@@ -2,11 +2,10 @@
     Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
     Developed by http://tacticalbacon.de
 */
-class CfgWeapons
+class CfgCommands
 {
-    class GMG_F;
-    class GMG_40mm : GMG_F
-    {
-        magazines[] += {"TB_Smoke_Grenades"};
+    allowedHTMLLoadURIs[] += {
+        "http://tacticalbacon.de*",
+        "https://tacticalbacon.de*"
     };
 };
