+++ addons/config/configs/CfgEditorCategories.hpp
@@ -4,5 +4,8 @@
 */
 class CfgEditorCategories
 {
-
+    class EdCat_TB_MainCat
+    {
+        displayName = "#TB - MainStuff";
+    };
 };
