+++ addons/config/configs/CfgEditorSubcategories.hpp
@@ -4,5 +4,18 @@
 */
 class CfgEditorSubcategories
 {
+    class EdSubcat_TB_Spezial
+    {
+        displayName = "Spezial";
+    };
 
+    class EdSubcat_TB_FOBs
+    {
+        displayName = "FOBs";
+    };
+
+    class EdSubcat_TB_BriefingBereich
+    {
+        displayName = "BriefingBereich";
+    };
 };
