+++ addons/config/configs/CfgFactionClasses.hpp
@@ -4,5 +4,9 @@
 */
 class CfgFactionClasses
 {
-
+    class NO_CATEGORY;
+    class TB_categorie: NO_CATEGORY
+    {
+        displayName = "#TacticalBacon";
+    };
 };
