+++ addons/config/configs/CfgMagazineWells.hpp
@@ -4,5 +4,8 @@
 */
 class CfgMagazineWells
 {
-
+    class UGL_40x36
+    {
+        ADDON[] = {"3Rnd_Smoke_Grenade_shell_precise"};
+    };
 };
