--- addons/config/configs/CfgUnitInsignia.hpp
@@ -1,8 +1,0 @@
-/*
-    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
-    Developed by http://tacticalbacon.de
-*/
-class CfgUnitInsignia
-{
-
-};
