+++ addons/config/configs/RscMapControl.hpp
@@ -1,0 +1,22 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+// TODO: Test, setzt das ACE3 nicht schon?!? und was sind Vanilla Werte?
+// ACE MircoDagr: https://github.com/acemod/ACE3/blob/e5a15d200f44df5fccc0bc5575d18d80b35538dd/addons/microdagr/gui_controls.hpp#L48
+// ACE Maps: https://github.com/acemod/ACE3/blob/e5a15d200f44df5fccc0bc5575d18d80b35538dd/addons/map/config.cpp#L39
+class RscMapControl
+{
+    maxSatelliteAlpha = 0.5; // 0.85;
+    colorBackground[] = {0.89999998,0.89999998,0.89999998,1}; // {0.969,0.957,0.949,1};
+    colorCountlines[] = {0.64999998,0.52999997,0.30000001,1}; // {0.572,0.354,0.188,0.25};
+    colorMainCountlines[] = {0.85000002,0,0}; // {0.572,0.354,0.188,0.5};
+    colorLevels[] = {0,0,0,1}; // {0.286,0.177,0.094,0.5};
+    sizeExLevel = 0.029999999; // 0.02;
+    colorTracks[] = {1,0,0,1}; // {0.84,0.76,0.65,0.15};
+    colorTracksFill[] = {1,1,0,1}; // {0.84,0.76,0.65,1};
+    colorRoads[] = {0,0,0,1}; // {0.7,0.7,0.7,1};
+    colorRoadsFill[] = {1,1,0,1}; // {1,1,1,1};
+    colorMainRoads[] = {0,0,0,1}; // {0.9,0.5,0.3,1};
+    colorMainRoadsFill[] = {1,0.60000002,0.40000001,1}; // {1,0.6,0.4,1};
+};
