+++ addons/config/configs/cfgScriptPaths.hpp
