+++ addons/main/config.cpp
@@ -141,14 +141,15 @@ class CfgPatches
             "ace_medical",
             "ace_medical_ai",
             "ace_medical_blood",
-            /*"ace_medical_damage", // NEW Ace Version
+            "ace_medical_damage",
             "ace_medical_engine",
             "ace_medical_feedback",
             "ace_medical_gui",
             "ace_medical_statemachine",
             "ace_medical_status",
             "ace_medical_treatment",
-            "ace_medical_vitals",*/
+            "ace_medical_vitals",
+            "ace_metis",
             "ace_microdagr",
             "ace_minedetector",
             "ace_missileguidance",
@@ -208,51 +209,12 @@ class CfgPatches
 };
 
 // Configs
-#include "configs\Animations.hpp"
-#include "configs\CfgAmmo.hpp"
-#include "configs\CfgEditorCategories.hpp"
-#include "configs\CfgEditorSubcategories.hpp"
-#include "configs\CfgFactionClasses.hpp"
-#include "configs\CfgMagazines.hpp"
-#include "configs\CfgMagazineWells.hpp"
 #include "configs\CfgMods.hpp"
 #include "configs\CfgSettings.hpp"
 #include "configs\CfgUnitInsignia.hpp"
 #include "configs\CfgVehicles.hpp"
 #include "configs\ExtendedEventHandlers.hpp"
-#include "configs\CBA_Stuff.hpp"
-#include "configs\CfgMisc.hpp"
-#include "configs\ACE_detector.hpp"
-#include "configs\CfgSounds.hpp"
-#include "configs\CfgWeapons.hpp"
 
 // Other config stuff
 #include "gcam\gcam_component.hpp"
 #include "configs\TB_RscHTML.hpp"
-
-// TODO: Test, setzt das ACE3 nicht schon?!? und was sind Vanilla Werte?
-// ACE MircoDagr: https://github.com/acemod/ACE3/blob/e5a15d200f44df5fccc0bc5575d18d80b35538dd/addons/microdagr/gui_controls.hpp#L48
-// ACE Maps: https://github.com/acemod/ACE3/blob/e5a15d200f44df5fccc0bc5575d18d80b35538dd/addons/map/config.cpp#L39
-class RscMapControl
-{
-    maxSatelliteAlpha = 0.5; // 0.85;
-    colorBackground[] = {0.89999998,0.89999998,0.89999998,1}; // {0.969,0.957,0.949,1};
-    colorCountlines[] = {0.64999998,0.52999997,0.30000001,1}; // {0.572,0.354,0.188,0.25};
-    colorMainCountlines[] = {0.85000002,0,0}; // {0.572,0.354,0.188,0.5};
-    colorLevels[] = {0,0,0,1}; // {0.286,0.177,0.094,0.5};
-    sizeExLevel = 0.029999999; // 0.02;
-    colorTracks[] = {1,0,0,1}; // {0.84,0.76,0.65,0.15};
-    colorTracksFill[] = {1,1,0,1}; // {0.84,0.76,0.65,1};
-    colorRoads[] = {0,0,0,1}; // {0.7,0.7,0.7,1};
-    colorRoadsFill[] = {1,1,0,1}; // {1,1,1,1};
-    colorMainRoads[] = {0,0,0,1}; // {0.9,0.5,0.3,1};
-    colorMainRoadsFill[] = {1,0.60000002,0.40000001,1}; // {1,0.6,0.4,1};
-};
-
-class CfgCommands
-{
-    allowedHTMLLoadURIs[] += {
-        "http://tacticalbacon.de*",
-        "https://tacticalbacon.de*"
-    };
-};
