--- addons/main/configs/CfgEditorSubcategories.hpp
@@ -1,17 +1,0 @@
-class CfgEditorSubcategories
-{
-    class EdSubcat_TB_Spezial
-    {
-        displayName = "Spezial";
-    };
-    
-    class EdSubcat_TB_FOBs
-    {
-        displayName = "FOBs";
-    };
-    
-    class EdSubcat_TB_BriefingBereich
-    {
-        displayName = "BriefingBereich";
-    };
-};
