--- addons/main/configs/CfgMagazineWells.hpp
@@ -1,10 +1,0 @@
-class CfgMagazineWells
-{
-   class UGL_40x36
-   {
-       TB_magazines_main[] =
-       {
-            "3Rnd_Smoke_Grenade_shell_precise"
-       };
-   };
-};
