--- addons/main/configs/CfgSounds.hpp
@@ -1,34 +1,0 @@
-/*
-    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
-    Developed by http://tacticalbacon.de
-*/
-class CfgSounds
-{
-    class ace_detector_1
-    {
-        name = "ace_detector_1";
-        sound[] = {"\z\ace\addons\minedetector\sounds\metal_detector.wss", "db+6", 1};
-        titles[] = {};
-    };
-
-    class ace_detector_2
-    {
-        name = "ace_detector_2";
-        sound[] = {"\z\ace\addons\minedetector\sounds\metal_detector.wss", "db+6", 0.9};
-        titles[] = {};
-    };
-
-    class ace_detector_3
-    {
-        name = "ace_detector_3";
-        sound[] = {"\z\ace\addons\minedetector\sounds\metal_detector.wss", "db+6", 0.8};
-        titles[] = {};
-    };
-
-    class ace_detector_4
-    {
-        name = "ace_detector_4";
-        sound[] = {"\z\ace\addons\minedetector\sounds\metal_detector.wss", "db+6", 0.7};
-        titles[] = {};
-    };
-};
